+++ example/ClassComplianceValidatorExample.php
@@ -2,21 +2,18 @@
 
 require __DIR__.'/../vendor/autoload.php';
 
-use LDL\Validators\Config\ClassComplianceValidatorConfig;
 use LDL\Validators\ClassComplianceValidator;
+use LDL\Validators\Collection\ValidatorCollection;
 use LDL\Validators\StringEqualsValidator;
 use LDL\Validators\Exception\ValidatorException;
 use LDL\Validators\RegexValidator;
-use LDL\Validators\Config\RegexValidatorConfig;
 
 echo "Create class compliance validator from config\n";
 echo "Set config class: 'RegexValidator'\n";
 
-$config = ClassComplianceValidatorConfig::fromArray([
+$validator = ClassComplianceValidator::fromConfig([
     'class' => RegexValidator::class
 ]);
-
-$validator = ClassComplianceValidator::fromConfig($config);
 
 echo "Validate integer, exception must be thrown\n";
 
@@ -26,10 +23,10 @@ try{
     echo "EXCEPTION: {$e->getMessage()}\n";
 }
 
-echo "Validate regex config, exception must be thrown\n";
+echo "Validate validator collection, exception must be thrown\n";
 
 try{
-    $validator->validate(new RegexValidatorConfig('#[0-9]+#'));
+    $validator->validate(new ValidatorCollection());
 }catch(ValidatorException $e){
     echo "EXCEPTION: {$e->getMessage()}\n";
 }