+++ example/RegexValidatorExample.php
@@ -3,17 +3,14 @@
 require __DIR__.'/../vendor/autoload.php';
 
 use LDL\Validators\RegexValidator;
-use LDL\Validators\Config\RegexValidatorConfig;
 use LDL\Validators\Exception\ValidatorException;
 
 echo "Create regex validator from config\n";
 echo "Set config regex: '#[0-9]+#'\n";
 
-$config = RegexValidatorConfig::fromArray([
+$validator = RegexValidator::fromConfig([
     'regex' => '#[0-9]+#'
 ]);
-
-$validator = RegexValidator::fromConfig($config);
 
 echo "Validate: 'string', exception must be thrown\n";